[dv] Retarget forced signal for prim_sparse_fsm_flop_if
Because the prim hierarchy has changed, reusing the path would target
the flop directly for forcing. This meant that the forcing would delay
u_state_flop from snapping back to the FsmError state that should've
been held by the flop. Instead, the original path caused the flop in the
new hierarchy to be targeted *directly*, delaying update until the next
posedge, instead of the negedge when forcing is released.

Retarget forcing to prim_sparse_fsm_flop's state_o output.

Signed-off-by: Alexander Williams <awill@opentitan.org>