[csrng/dv] Exclude AES core from coverage analysis
This commit excludes the AES cipher core from the coverage analysis.
The exlusions file had to be revised since our coverage plummeted even
though the coverage for CSRNG core was close to 100%.
In the past we excluded a lot of paths from the AES core individually.
However, excluding the AES core all together might be a more honest
and transparent approach, since we now also exclude the paths where
we have 100% coverage inside the AES core.

Signed-off-by: Hakim Filali <hfilali@lowrisc.org>