[i2c,dv] Switch sequence handle to base vseq for factory overrides
This change allows derived vseqs to override the stimulus sequence run on the
agent by using the factory mechanism.

Also, change the variable name to 'm_i2c_controller_seq', as I2C uses the term
controller where OpenTitan uses host (i.e. the bus initiator), and trying to
maintain consistency in terms should hopefully make this whole thing slightly
less error prone. In this case, the controller_seq runs on the I2C_Agent, which
drives the agent to behave as an I2C Controller for a single transaction.

Signed-off-by: Harry Callahan <hcallahan@lowrisc.org>