[usbdev] Increase robustness of packet reception
End of packet must always be signaled to the packet state
machines even if the packet has been declared invalid, because
otherwise the OUT side state machine will wait indefinitely for
the end of a DATA packet.
Modify end of packet detection so that it does not depend upon
the validity of the SYNC signal.
Additional state is introduced to track that reception is still
mid-packet but the packet is being ignored due to a PRE token or
a bit stuffing violation. This allows the logic to avoid false
restarts in response to ensuing packet data that looks like a
full speed SYNC signal.

Signed-off-by: Adrian Lees <a.lees@lowrisc.org>