[manuf] refactor cert size reading
This refactors the cert size reading/decoding logic from the
personalization binary into separate libraries. This is done so the code
may be reused by the ROM_EXT, or other owner firmware, that may try to
update or read certificates from flash info pages (particularly if multiple
certificates are stored on the same page).

Signed-off-by: Tim Trippel <ttrippel@google.com>