[rv_dm,dv] Properly implement jtag_dmi_cmderr_busy testpoint
The existing implementation wasn't very sensible: it had a bunch of
magic hex constants and checked all the operations without clearing
their effect (so it was really just checking the first one).

It also didn't clean up after itself at the end of the test, which
meant that it rather breaks the stress_all test (because there's an
abstract command running when some other vseq tries to do something).

Tidy this up.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>