[rv_dv,dv] Rewrite rv_dm_ndmreset_req_vseq
This should hopefully be a bit easier to understand now, but is doing
essentially the same operations as before.

I've also dropped the repetition inside the sequence: there doesn't
seem much benefit to the repeats, and that sort of "do it repeatedly"
behaviour is really implemented in the stress_all sequence.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>