[rv_dm,dv] Slightly weaken ndmreset_req testpoint
Part of the description is that we write some known value to some
registers and then read those back after the ndmreset and check the
value is unchanged.

The description was talking about three different RALs, one of which
is the debug register interface. But we don't actually have any
appropriate registers there (we need a "scratch" register that we can
write with any old value). The existing test didn't actually touch
that interface and I think that's probably reasonable. So stop
requiring that we do.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>