[rv_dm,dv] Tidy up lc_hw_debug_en control
This gets driven by the base vseq and the value of the signal is
constrained in pretty much every subclass, which seems a little silly.
Move the duplicated code to the base vseq.

The reason I noticed this at all is that it *wasn't* duplicated in
rv_dm_stress_all_vseq, causing silly test failures. This fixes them.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>