[rom] Improve boot failure error reporting
Improve the error reporting when the ROM fails to boot one of the
slots. Often, one of the slots has an image and the other is empty. If
there is an error with the image, the ROM will try the empty slot and
then report the "Bad Identifier" error from the empty slot rather than
the error associated with the bad image.

Addresses #23724

Signed-off-by: Chris Frantz <cfrantz@google.com>