[rv_dm,dv] Avoid writing to command in rv_dm_jtag_dtm_idle_hint_vseq
Writing a command leaves things in an untidy state for following
sequences: we've either started a command and need to finish it, or
we've generated an error because of the invalid command and need to
clear the error.

This testpoint is about doing back-to-back DMI accesses, so there's no
real need to think about abstract commands at all. Don't!

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>