[kmac, dv] Test KMAC capacity zeroing
Update KMAC DIF and testutils, so it can also
read the capacity part. Normally the capacity
of Keccak state is not read as part of cryptographic
operation but we need this for testing.

This commit also updates the sideload kmac
test to check that:
i) For sideloaded KMAC, the capacity is zeroized
by KMAC HWIP.
ii) For SW-keyed KMAC, the capacity can be read
by SW.

See #17759.

Signed-off-by: Fatih Balli <fatihballi@google.com>