[dv,usbdev] Exercise all bus configurations
There are three bus configuration options, leading to eight
possible types of bus. Ensure that all are tested.
Modify the usb20_driver/monitor to be aware of pin-flipping when
the TX side is using D/SE0 because this does invert the sense
of the 'd' signal, since it is effectively driving DN now.

Signed-off-by: Adrian Lees <a.lees@lowrisc.org>