[entropy_src/rtl] Delay the main_sm_done signal
Due to the relocation of the 1-4-bit packer FIFO, the main SM done
signal arrived one cycle too early in the single lane mode. This
caused the SHA3 process_i signal to go high one cycle too early.
This in turn caused the issue that the last word from the precon
FIFO is not absorbed properly by SHA3.
This commit fixes this by delaying the main_sm_done signal by 1 cycle
in the single lane mode. The ht_error signal has to be delayed by 1
cycle as well to guarantee that the error signal arrives at the same
time as the main_sm_done signal.

Signed-off-by: Hakim Filali <hfilali@lowrisc.org>