[i2c,testutils] Fix i2c_testutils_wait_transaction_finish
This function was not updated when the RTL was changed to halt the
FSM on an unexpected NAK, and therefore would spin together. Since
previously the FSM would run and drain the FIFO even when a NAK
is received, the semantic of this function has to change. With this
commit, it now returns an OK_STATUS indicating whether a halt event
was detected.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>