[i2c,host_irq] Fix test
The test was previously using i2c_testutils_read to issue the read
and then checks for halt events. But the test utils were updated
to handle halt events at some points so when the test tries to read
them, they have already been erased. To ensure a reliability, use
a low-level DIF function to trigger the read instead.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>