[bazel] fix bug in FGPA tagging scheme
The "cw310" tag was getting applied to all "fpga_params()"
instantiations, regardless if the targeted FPGA was the CW310 or not.
This applies the correct FPGA target tag using the "_hacky_tags()"
helper method instead.

This was causing issues as we use the tags to filter which tests to run
on which FPGA boards in CI.

Signed-off-by: Tim Trippel <ttrippel@google.com>