[rv_dm,dv]: Implement sec_cm_lc_hw_debug_en_intersig_mubi testpoint
Most of the testpoint is implemented in new rv_dm_debug_disabled_vseq,
which is new. Here, we want to check the multi-bit encoding of the
lc_hw_debug_en_i input. We drive that it to some value other than On
and do an operation that should have no effect.

One check that was already implemented was that SBA access is not
possible: the JTAG DMI interface can't generate TL transactions. This
is implemented in rv_dm_sba_debug_disabled.

As well as implementing things, I've expanded out the description of
the testpoint to explain what we're doing.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>