[entropy_src/dv] Remove unimportant cover group bins
This commit removes some bins from crosses that are either not important
or even an illegal combination of settings.
First some bins are removed where some MuBi values are set to invalid.
Second some bins are removed that set the threshold scope to true when
the single bit mode is active. In the single lane mode it does not make
sense to evaluate all 4 lanes together.

Signed-off-by: Hakim Filali <hfilali@lowrisc.org>