[rv_dm,dv] Support "late enable through register" mechanism
The logic for whether debug is enabled in the dut looks like:

  if ((otp_dis_rv_dm_late_debug_i is mubi true) ||
      (the late_debug_enable register is mubi true)) {
    enabled = lc_hw_debug_en_i;
  } else {
    enabled = lc_dft_en_i;
  }

Before this commit, we always set otp_dis_rv_dm_late_debug_i to true.
This commit makes it possible to depend on the register because
otp_dis_rv_dm_late_debug_i doesn't have the "true" value.

We still aren't handling the situation where neither is a mubi true
value, but that will come in a following commit.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>