[i2c,fram_test] Reduce expected throughput in fast plus mode
Due to the updates to the timing to guarante that clock stretching works,
fastplus mode may not reach 1Mbit/s, especially on the FPGA. The simplest
fix is to slightly lower the expected speed. An alternative would be to
compute the expected speed from the actually computed i2c timing.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>