[rv_dm,dv] Config lc_hw_debug_en as a bit, not a 4-bit variable
The fact that we encode this with a multi-bit value is a "driver
issue" and probably doesn't belong in the derived sequences. This
commit changes the "enabled/disabled" state to be a 1-bit variable and
moves the randomisation for "disabled" values to the new
upd_lc_hw_debug_en() function.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>