[Cadence FCOV flow] Cadence FCOV ditched if per_instance=1' not added
Cadence coverage flow has decided not to merge FCOV from covergroups
which don't have 'option.per_instance=1' regardless of whether this
covergroups are embedded in a class or instantiated.

In order to continue gathering FCOV in the way we intend to, I suggest
configuring Cadence tool with a flag so we still collect FCOV from
covergroups when per_instance=0.

Signed-off-by: Antonio Martinez Zambrana <amz@lowrisc.org>