[lc_ctrl,dv] Add nonzero stagger lengths between rma signals
This uses the "delay_lens" argument that the previous commit added to
cfg.lc_ctrl_vif and randomizes those delay lengths in the vseq.

The only surprising behaviour here is that we *don't* randomise the
timing when we're implementing the flash_rma_error_rsp behaviour.
There, we set max_delay_cycles to zero so that the
scoreboard-embedded-in-a-vseq in the file tracks things correctly.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>