[chip-test/sival] Fix csrng_edn_concurrency_reduced_freq
Inspecting the waves revealed that the reduced_freq variant of this test
sometimes didn't make it across the initial EDN and CSRNG re-config
section. The reason is that the RNG inside AST produces raw entropy
at a lower rate while the entropy consuming part of AST still consumes
a reasonable amount of entropy. As a result and depending on the EDN
config, the esfinal FIFO would never fill up and the reconfiguration
section would never finish.

To fix this, this commit adds a timeout for the esfinal FIFO: If the
FIFO hasn't filled up within 32 ms (under normal conditions, this should
suffice to fill the FIFO easily), we proceed without a full FIFO.

This fixes lowRISC/OpenTitan#23606.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>