[opentitanlib] Don't set the baudrate on pseudo devices
The verilator UART implementation uses pseudo-devices to facilitate UART
communication between the simulation and outside processes.  The
simulation and the test-harness may not agree on the baud rate and,
since there is no real UART involved, this should not be an error.
(regardless of your opinion on whether or not it should be an error, the
rust serialport crate treats a disagreement in rates as not only an
error, but a program-ending _panic_ event).

To allow tests to execute, don't get/set the baudrate for
pseudo-devices.

Signed-off-by: Chris Frantz <cfrantz@google.com>