[chip-test/sival] Check ENTRORPY_SRC_POSTHT_ENTROPY_DROP_ALERT_BIT
When running in Firmware Override: Observe mode, this bit should be
checked. If it's asserted, this means that health tested entropy bits
have been dropped before the postht FIFO due to conditioner back
pressure. This means entropy observed via observe FIFO may be
non-contiguous.

This is related to lowRISC/OpenTitan#21855.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>