[entropy_src/dv] Reset CSRNG IF after CS AES Halt IF becoming idle
Previously, the CSRNG interface got occasionally reset at the end
of entropy_src_rng_vseq while the CS AES Halt interface was still busy,
i.e., while the conditioner is active. Since the CS AES Halt interface
is reset together with the CSRNG interface, this led the agent to drop
the ack signal while the req (driven by the conditioner) was still high.
As a result we would occasionally see an SVA firing which checks that
the conditioner is never active without a valid handshake on the CS AES
Halt interface.

This commit solves the issue by waiting for the CS AES Halt interface to
first become idle before resetting the CSRNG interface.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>