[otbn,dv] Fix a bug in random instruction generator
This was supposed to be picking a branch target that would be "bad"
for various reasons. In the case that I tweaked, the problem was just
that it wasn't a multiple of 4.

This target address is encoded as a 12-bit signed value of 16-bit
words, so needs constraining to be within 4k bytes of the current PC.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>