[rv_core_ibex,dv] Fault lockstep enable signal
This DV only commit extends the chip_sw_rv_core_ibex_lockstep_glitch
test in such a way that the lockstep enable signal is additionally
faulted.

The Ibex lockstep tolerates a fault into the enable_cmp_q due to the
MuBi encoding and how the lockstep enable RTL is written.

This commit randomly inverts one bit in the MuBi encoded enable signal.
The test should behave as before - an alert should be triggered.

Closes #23572.

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>