[opentitantool] Declare additional aliases of HyperDebug pins
HyperDebug has no flexibility on which pins to use for SPI1, SPI2,
UARTs, etc.  This CL declares a number of alises, that allow referring
to pins such as "UART3_TX", which may be convenient from tests which
want to use e.g. `GpioMonitoring` for logic analyzer functionality on
pins used for UART, SPI or I2C.

Change-Id: I58c3af6bf85e31de067e6813969f58de4347894b
Signed-off-by: Jes B. Klinke <jbk@chromium.org>