[flash_ctrl/dv] Add sec_cm_mem_addr_infection test
This commit adds a new test to verify the MEM.ADDR_INFECTION countermeasure.
The test issues a host read request to the flash controller, manipulates
the address inside the flash_phy_rd module, and checks, whether the TL-UL
data integrity error is received. This error is expected as the
MEM.ADDR_INFECTION mitigation should infect the data with the faulty address
leading to an integrity error.

Closes #23898

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>