[i2c,dv] Add new types for i2c TLM modelling
- i2c_pkg::rw_e and i2c_pkg::acknack_e are used for readability.
- When working with the transfer-direction bit at the end of the first byte of
  every transfer (rw_e = READ/WRITE)
- The ACK enum (acknack_e = ACK/NACK) represent the acknowledgement bit at the
  end of every data byte in a normal transfer.

- The classes i2c_acqdata_item and i2c_fdata_item are added, which model data
words written to the corresponding control FIFOs. Previously, fields from each
of these fifos were dropped into the i2c_item as class members, and were used
for a number of subtly-different functions across the agent and env. These are
not yet completely exfiltrated into the specific items. For example, there is still a
'start' and 'stop' field in the i2c_item, which overlap with the same name
fields in the FMTFIFO/fdata word, and are sometimes used as a) capturing bits
written to the register and b) to represent a start or stop condition being
associated a transfer or byte. This will further be improved in the future.

Signed-off-by: Harry Callahan <hcallahan@lowrisc.org>