[i2c,dv] Improve comments and code structure in i2c_target_base_seq
The agent sequences will require further work to cleanup, but this is a start.
This file in particular is in desperate need for a rename, as the sequence
drives the agent as an I2C Controller, not a target! In a number of places, we
construct sequences with instance names that appear contradictory...
```
    i2c_target_base_seq m_i2c_host_seq;
```

I suspect it was probably named for the mode it was involved in
testing (DUT-Target Mode), but we should really be naming things related to
their own function. The agent could standalone among many other Controllers and
Targets on the bus after all.

Signed-off-by: Harry Callahan <hcallahan@lowrisc.org>