[i2c,dv] Consolidate I2C item to model at transfer-level, use monitor output for expectations
This large commit fixes a number of fundamental flaws and confusing design
decisions around the way we model TLM transactions using the i2c_item.

The major flaw this change initally set out to fix was that scoreboard
expectations were being formed by passing the generated stimulus from the
virtual sequences directly to the model/scoreboard. This is against UVM design
patterns, and severely limits the ability to predict behaviour when the effect
of the driver are fundamentally bypassed and unpredictable. In I2C, the receiver
in any transaction can alter the progression of the traffic by choosing to ACK
or to stretch SCL. In our I2C environment, where both endpoints are
operated/tested as both Controllers and Targets, this means any reactivity
in the Agent cannot be predictable, and sequences are therefore reduced to only
a very small subset of behaviour. This problem was partially circumvented by a
slew of configuration variables used to hold bits of state and signal events to other
components, but this is unclean and inconvenient.

To fix this problem, another intermediate issue needed to be addressed. The
i2c_item was grossly overloaded in the way it was used to represent different
things to different people across the testbench and agent. For read transfers,
the seq_item was used to capture the entire transfer (data_q[$] held all the
bytes of data), but for writes we would capture one item per entry in the
ACQFIFO CSR to signal an I2C transaction was taking place. Different fields in
the item were used for each use-case, and some were re-used with different
meanings for each case.
> There is also a third use of the i2c_item, again using different fields (with
> some overlap), which is to control the i2c_driver in discrete steps depending
> on the value of a control enum variable 'drv_type'. This use case remains, but
> should definitely be removed in the future.
To model at a transfer level, new fields 'rstart_front' and 'rstart_back' were
added to the item, which indicate if the start or the end of the transaction was
signalled using RSTARTS instead of STARTs and STOPs. The previous 'rstart' field
would be ambiguous when used for modelling at a transfer-level.

Additionally, this commit also importantly changes the way mis-addressed
transfers are captured and checked. Previously, the i2c_agent contained copies
of the DUT's configured Target-Mode addresses, and used these to change the way
it captured observations and generated stimulus based on if the address given
was known ahead of time to match the DUT. This is clearly a badly-designed
dependence between the modules, and by removing this pre-filtering of
transaction items, we now generate and check items which do not match the
targets address. These items are effectively the I2C-Controller shouting into
the void, which is a valid i2c transfer in terms of signalling, but in a single
controller, single target system, a wholly uninteresting one.

In short, this commit achieves the following:
- Reworks the i2c_item to model full 'i2c_transfers', and updates the
  i2c_monitor to generate seq_items at this level of abstraction, for both reads
  and writes.
- Change the reference_model/scoreboard to gather observations in this new
  format from the monitor only (connections to the vseqr are now removed)
- The i2c_monitor now generates valid transfer items for mis-addressed stimulus,
  easily observed by checking if the ITEM_T.addr_ack field is equal to i2c_pkg::NACK.
- Replaces the stimulus generation routines in the i2c_base_vseq with new,
  concise versions that no longer need to generate transaction items to send to
  the model/scoreboard. The previous code here was verbose, unclear and
  difficult to modify.
  - create_txn()/fetch_txn() now replaced by a single call to
  generate_agent_controller_stimulus(), which makes better use of types,
  targeted subroutines and randomization to achieve the same goal in a clear and
  simpler way.
- Renames variables across the env/agent cfgs, moved variables from the
  class-scope into task/function scope where possible, and used user-defined
  types where appropriate to aid readability.

Signed-off-by: Harry Callahan <hcallahan@lowrisc.org>