[entropy_src] Use `es_delayed_enable` for SVAs tracking entropy flow
Previously, these SVAs were using the undelayed `mubi_es_enable` signal
to decide whether the module was on or off which led to corner case
issues in several tests. Using `es_delayed_enable` makes more sense as
the clearing signals for the FSMs, counters and FIFOs depend on this
signal instead of `mubi_es_enable`.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>