[kmac,dv] Remove `err_processed` from shadow reg DV logic
PR lowRISC/opentitan#22794 moved the `err_processed` field from
the `CFG_SHADOWED` shadow register to the `CMD` register.

However, the `shadow_reg_wr` function was not adapted. This function
assumes that writing to the `CFG_SHADOWED` reg clears the update error
field. As this now can only be done with the `CMD` register, this is
not true anymore. Hence, this commit removes this assumption.

Closes lowRISC/opentitan#23083.

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>