[otbn,dv] Properly track partial wipes in model
This is to correctly handle a situation where we decide to jump to the
locked state in the middle of a secure wipe. Before this change, we
would print out

 ..., U, U, U, U
  (jump to locked state)
 V, STALL, STALL, STALL, ...

and the trace checker would get upset that the model sent a V (meaning
that a wipe had completed) and then a STALL without any corresponding
event from the RTL side.

With this change, we get ..., U, U, U, U, STALL, STALL, ... and avoid
the error.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>