[otbn,dv] Lock on URND completion when wiping
In the RTL, this is considered a spurious URND ack. You can see it if
there is an RMA shortly after reset (after the URND request goes out,
but before the seed comes back). OTBN obeys the RMA and jumps to a
wiping state, but then locks when the URND ack comes back: "we're in
the middle of a secure wipe, so don't think we've requested anything".

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>