[dv,usbdev] Monitor improvements
Simplify the collection of DATA packets.
Resolve the confusion about item constructors and make
them more appropriate for use with the UVM factory; it is
not possible to provide alternative parameters to the new()
constructor via the factory.
Ensure that the monitor listens to d/se0 only when the DUT
is talking, since the driver cannot produce traffic on those
signal lines; they are not shared/tri-stated.

Signed-off-by: Adrian Lees <a.lees@lowrisc.org>