[i2c,dv] Don't clear the item state for each new monitor transfer
This was an artifact of how the monitor previously captured data, by re-using
the same sequence item over and over again to track the current transfer state,
and then cloning it whenever it wished to publish an item.

Now that we are exporting an in-progress transfer item that is meant to remain
alive and valid for a longer period of time, we do not wish to clear the state
anymore.

Signed-off-by: Harry Callahan <hcallahan@lowrisc.org>