[i2v,dv] Re-implement i2c_base_seq using the 'in-progress' xfer_item
This particularly affects Agent-Target operation, where we need to be reactive
to the different components of the transfer as they are received from the
I2C-Controller. In particular, we want to hook the following parts of the
transfer:
1) The end of the addr+dir byte (Should we ACK or NACK this address?)
2) The end of each write data byte (Should we ACK or NACK this received byte?)
3) The start of each read data byte (We need to drive the byte)

This change re-works the base class to provide a simple hook-based system for
child sequences to implement more-specific drive behaviours.

Signed-off-by: Harry Callahan <hcallahan@lowrisc.org>