[otbn,dv] Tidy up the timing for when we detect invalid RMA requests
A bogus RMA request only matters at specific states in the start/stop
controller FSM. Change the model so that it similarly only looks for
bad values at those times. Annoyingly, this takes effect "just after"
we finish the secure wipe, so we need to model it with a new
delayed_lock variable.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>