[otbn,dv] Merge WIPING_GOOD, WIPING_BAD
I split these apart when I first handled secure wipe. It seemed like a
good idea at the time! But it's not so great, especially if I want to
split out the "waiting for URND seed" state (and don't want to
duplicate that). Handle the state we're wiping towards with a separate
flag.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>