[opentitanlib] preprocess the manifest before signing
The manifest usage_constraints field includes values which could be
ignored by the signature validator, depending on the selector_bits
setting.

This patch makes sure that the signer behaves the same, i.e. replaces
the values disabled by the selector with the default value.

Tested by successfully starting the Ti50 image on the Cw310 FPGA
board.

Signed-off-by: Vadim Bendebury <vbendeb@google.com>