[i2c,dv] Integrate i2c monitor data into reference model predictions
This is part of a larger push to improve the i2c_reference_model to start making
better and more granular predictions.

This commit achieves the following:
- Connects the i2c_monitor's target-mode transfer ports and in-progress ports
through to the reference model.
- Creates a local queue of predicted controller-mode transfers based on CSR
access stimulus that defines the transfers. Rather than pushing these items
directly to the scoreboard, we need to keep an in-order list to fuse
with bus observations from the i2c monitor.
  - Create a separate queue of both read and write transfers (prev_item[$])
  together, as determination of the next transfer always depends on the
  ending conditions of the previous one, no matter if it was a read or a write.
- Fuses observed ACK/NACK bits from the bus with the predictions from the CSR
accesses. This is not yet particularly interesting, but will be crucial for
better modelling mid-transfer inputs in the future.

Signed-off-by: Harry Callahan <hcallahan@lowrisc.org>