[rv_dm,dv] Avoid some scoreboard behaviour in sba_access_monitor
This wasn't really in the right place: something has gone wrong if the
design can cause a monitor to die!

It's completely wrong in the rv_dm setting. If debug is disabled then
errors won't appear on the JTAG interface, so we don't see the error
flags that the monitor was expecting.

Stop requiring them.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>