[otbn,dv] Remove short-circuit from start() in sim.py
This was added in commit a491a03. I'm not completely convinced that
the change was correct, and it isn't really explained in the commit
message. (And it did break something else for me). I will remove the
short-circuit for now. If it turns out to matter, I'll put it back in
a follow-up commit, with extra notes that explain what it's doing.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>