[signing] Support ECDSA signing with hsmtool
Support ECDSA signing with `hsmtool`.

Currently, in order to create an ECDSA signature with hsmtool, you must
have keys defined in a `keyset`.  Longer-term, we should eliminate
`keyset`s in favor of the `KeyInfo` provider introduced in #22155.
This change enables signing with an ECDSA keyset to test hsmtool signing
flows with the ROM.

Signed-off-by: Chris Frantz <cfrantz@google.com>