[rom_e2e,dv] fix failing ROM E2E self hash test in DV
This fixes the ROM E2E self hash test in DV as it was failing due to
having an empty manifest.

Additionally, this moves the test to run in the PROD state by updating
the OTP image and signing keys. Since the test is now signed with the
PROD keys, it should boot in all LC states, if the LC state is changed
in the OTP image.

Signed-off-by: Tim Trippel <ttrippel@google.com>