[i2c,dv] Spawn i2c_agent base_seq driving hooks as background tasks
The way that the i2c_agent drives the interface reactively currently has a small
race condition. We watch the in-progress transfer field 'inp_xfer.state' to know
when the transfer has progressed to a new state, but sometimes the previous
driving hook returns on the same timestep as the state variable changes. This
causes the agent to miss the state-change and hence miss it's hook for when the
bus should be driven.

I think this architecture could be iterated on in the future, but for now a
solution is to schedule bus driving events into background processes and to not
await them to finish before going back to wait on the next state change. These
hooks should not overlap in practice. Since we do not change i2c_agent sequences
on-the-fly during testcases, these background threads can all complete in their
own time.

Signed-off-by: Harry Callahan <hcallahan@lowrisc.org>