[entropy_src/dv] Add prediction for the fw_ov_wr_fifo_full csr
This commit adds prediction for the fw_ov_wr_fifo_full csr to the
scoreboard. It also adds a read to the fw_ov_wr_fifo_full register
whenever a word is written to the fw_ov_wr_data register to actually
check if the returned values for the full signal match the prediction.

Signed-off-by: Hakim Filali <hfilali@lowrisc.org>