[rv_dm,dv] Use the "right enable" in *_jtag_dmi_debug_disabled_vseq
The code here was a bit broken and I clearly hadn't previously
understood the difference between "pinmux enabled" and "debug
enabled".

The *whole point* of this test is supposed to be that DMI transactions
get blocked when pinmux is not enabled. After the changes here, we
actually test that property!

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>