[bazel,ci] Process multiple outputs in target-location.sh
This commit switches target-location.sh from using `bazel outquery` to
`bazel cquery --output=files`, so that for the case of an
`opentitan_binary` software target with multiple potential outputs, we
can find any of them using the script. We ensure the execution root is
added to all output files rather than just the first, so that all
queried files are translated to absolute paths.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>