[dv,top_earlgrey] Fix pwrmgr_rstmgr_sva_if bind for top
The assertions pwrmgr_rstmgr_sva_if need to be disabled when either aon or
main_pok are inactive. It is more reliable to bind them to rstmgr for
top-level simulations. This PR does that and connects rst_slow_ni to
&rst_por_aon_n, since rst_por_aon_n is 2-bit wide (one per domain), and
captures the rstmgr behavior more accurately.

This also removes the rstmgr_unit_only* files for simplicity.

Fixes #23961

Signed-off-by: Guillermo Maturana <maturana@opentitan.org>