[docs,bazel] Correct FPGA bootstrapping in docs
Adds the `fpga_cw340` execution environment to the `hello_world` demo
example `opentitan_binary`, so that the demo can be built and run on the
CW340 like is advertised in the docs.

Documentation files are then updated to provide commands to properly
build and locate this binary in accordance with their current usage.
Also fixes a couple of semantic errors, and provides some extra
information and troublshooting.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>