[bazel] Ensure output files of outquery are unique
Resolves OpenTitan/Issue#23990. Some scripts rely on the fact that
outquery only returns a single file corresponding to some unique
software target, but this is not necessarily the case when an
`opentitan_test` is built using `build`, as the target is built in
multiple configurations, so `cquery` returns the same file multiple
times.

This is a temporary fix; a better solution in the future would be to
replace `--output=starlark` with `--output=files`, but this requires
some additional work.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>
(cherry picked from commit 041df33c9848b3ca053a4405000103ebadafd3a1)