[rv_dm,dv] Remove "TBO" comment from rv_dm testplan
This comment was quite reasonable: it's pretty hard to see how to
create other types of failed DMI transaction. The obvious approach of
accessing particular registers when an abstract command is running
doesn't work very easily because there's a very precise timing window
you have to hit.

But it's not clear why we care! This testpoint is all about making
sure that a failed DMI operation gets reported and that we can clear
that status again. The linked test implements that behaviour.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>