[chip-test,sival] Optimize ENTROPY_SRC config for ast_clk_rst_inputs
This test previously configured the ENTROPY_SRC in Firmware Override:
Extract and Insert mode and erroneously enabled the routing of
entropy to firmware (ENTROPY_DATA register). Also the entropy source
was reconfigured multiple times during the test.

This commit modifies the test to 1) use the Firmware Override: Observe
mode in which entropy is observable from the Observe FIFO but
also continues to flow through the pipeline to CSRNG and the EDNs, and
2) to not re-configure the entropy source but just restart it to restart
the entropy re-collection.

In addition, the default_spinwait_timeout_ns value in the vseq is
is updated together with the run_timeout_mins to avoid random timeouts.
Also a check that the observe FIFO is empty when we come out of deep
sleep is removed. This check doesn't make sense as after deep sleep, a
reset is performed which empties the observe FIFO anyway.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>