[i2c,dv] Fixup i2c_target_smoke_vseq handling of agent_cfg.got_stop
This is necessary for the upcoming fixes to the target_stress_all test.
Without clearing this variable on all iterations of the 'num_trans' loop, we run
into problems when `num_trans == 1` and we are connecting multiple vseq's
together in the stress_all test.
This was painful to debug, and just re-inforces why these sort of flags in the
agent/env cfgs are really bad practice. They are used as test-long state, which
must be managed alongside the lifetime of any sequences. This is incredibly
error-prone, hard to debug, and exponentially increases the barrier to entry
when trying to grok the testbench architecture and system design.
As tracked in #23920, this will be removed in the future

Signed-off-by: Harry Callahan <hcallahan@lowrisc.org>