[i2c,dv] Fixup test failures in the i2c_target_hrst_vseq
This test falls squarely under the 'needs improvement' category, so this commit
takes the opportunity to remove & clarify some of its worse aspects while
getting things going again after previous changes caused it to regress.

- This test's name and description is confusing, to start with. I think 'hrst'
may have been an acronym for 'host-reset', but this test does not do that. It's
description talks about introducing 'glitches', but really it introduces
unexpected, non-protocol START and STOP conditions which cause a transfer to end
early. These aren't really 'glitches' at all, as it is valid to send a S or P at
any time.
- The test constrains stimulus to only use 'WRITE-transfers', and then points to
"glitches" in READ-transfer being covered by the i2c_glitch_vseq. As a result,
all of the code in this vseq relating to generating READ-transfer stimulus is
dead. Which is about 30% of the whole module. All this code has been removed,
and the existing implementation simplified to reflect this.
- The task "stop_target_interrupt_handler()" just duplicates the base class
implementation. Removed.
- The vseq tried to keep the scoreboard/env alive after introducing each
"glitch" by reaching out and clearing state in the i2c_agent's monitor and also
the env's scoreboard. This is deeply unpleasant, and added a whole bunch of code
to both of those modules just for this particular testcase. There is in-fact a
much easier way to do this, which is now possible after previous refactors to
make the i2c_item model a single i2c-transfer at a TLM level. An early START or
STOP condition just signals an early end-of-transfer, and we simply need to
ensure the monitor will end the transfer when it sees this condition. This logic
has already been added to most routines inside the monitor, but it was not added
to the routine which captures the address+direction byte, so it has been added
now. All of the sideband 'cfg.monitor_rst' mechanism has been removed, as has
the code which reaches in and artificially ends the transfer early. This is no
longer needed, now that we just properly model the S/P condition ending the
current transfer.
  - This also removes the need to mess around with 'skipping-starts' and other
  changes to stimulus, as the monitor will automatically start capturing a new
  transfer if the "glitch" stimulus event is a START (i.e. an RSTART in this context).

Signed-off-by: Harry Callahan <hcallahan@lowrisc.org>